services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "nutrI-plantA: servicio de autenticación y gestión de usuarios"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    # Validación y configuración
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "pydantic[email]==2.10.4",
    # Base de datos
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    # Migraciones
    "alembic==1.14.0",
    # Seguridad
    "pyjwt==2.10.1",
    "pwdlib[argon2]==0.2.1",
    # DI
    "dependency-injector==4.43.0",
    # Logging estructurado
    "structlog==24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
    "ruff==0.8.4",
    "mypy==1.14.0",
    "import-linter==2.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# ============================================================
# Ruff: linter rápido
# ============================================================
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long (delegado a formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]  # asserts permitidos en tests
"src/auth/api/**" = ["B008"]  # FastAPI usa Depends() en defaults por diseño
"src/auth/infrastructure/persistence/user_model.py" = ["N811"]  # SQLAlchemy UUID

# ============================================================
# Mypy: type checker estricto
# ============================================================
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false  # FastAPI usa decoradores no tipados

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ============================================================
# Pytest
# ============================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
]
markers = [
    "unit: tests unitarios sin dependencias externas",
    "integration: tests con BD u otros servicios",
]

[tool.coverage.run]
source = ["src/auth"]
omit = [
    "*/main.py",
    "*/__init__.py",
    "*/alembic/*",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false

# ============================================================
# Import-linter: reglas de arquitectura hexagonal
# ============================================================
[tool.importlinter]
root_package = "auth"

[[tool.importlinter.contracts]]
name = "domain no depende de infrastructure"
type = "forbidden"
source_modules = ["auth.domain"]
forbidden_modules = ["auth.infrastructure"]

[[tool.importlinter.contracts]]
name = "domain no depende de api"
type = "forbidden"
source_modules = ["auth.domain"]
forbidden_modules = ["auth.api"]

[[tool.importlinter.contracts]]
name = "infrastructure no depende de api"
type = "forbidden"
source_modules = ["auth.infrastructure"]
forbidden_modules = ["auth.api"]
